+++ core/unittest/server/test_web.cpp
@@ -53,6 +53,8 @@
 #include "server/DBWrapper.h"
 #include "utils/CommonUtil.h"
 
+#include "unittest/server/utils.h"
+
 static const char* TABLE_NAME = "test_web";
 static constexpr int64_t TABLE_DIM = 256;
 static constexpr int64_t INDEX_FILE_SIZE = 1024;
@@ -66,6 +68,7 @@ using OQueryParams = milvus::server::web::OQueryParams;
 using OChunkedBuffer = oatpp::data::stream::ChunkedBuffer;
 using OOutputStream = oatpp::data::stream::BufferOutputStream;
 using OFloat32 = milvus::server::web::OFloat32;
+using OInt64 = milvus::server::web::OInt64;
 template<class T>
 using OList = milvus::server::web::OList<T>;
 
@@ -86,11 +89,34 @@ RandomRowRecordDto(int64_t dim) {
     return row_record_dto;
 }
 
+OList<OInt64>::ObjectWrapper
+RandomBinRowRecordDto(int64_t dim) {
+    auto row_record_dto = OList<OInt64>::createShared();
+
+    std::default_random_engine e;
+    std::uniform_real_distribution<float> u(0, 255);
+    for (size_t i = 0; i < dim / 8; i++) {
+        row_record_dto->pushBack(static_cast<int64_t>(u(e)));
+    }
+
+    return row_record_dto;
+}
+
 OList<OList<OFloat32>::ObjectWrapper>::ObjectWrapper
 RandomRecordsDto(int64_t dim, int64_t num) {
     auto records_dto = OList<OList<OFloat32>::ObjectWrapper>::createShared();
     for (size_t i = 0; i < num; i++) {
         records_dto->pushBack(RandomRowRecordDto(dim));
+    }
+
+    return records_dto;
+}
+
+OList<OList<OInt64>::ObjectWrapper>::ObjectWrapper
+RandomBinRecordsDto(int64_t dim, int64_t num) {
+    auto records_dto = OList<OList<OInt64>::ObjectWrapper>::createShared();
+    for (size_t i = 0; i < num; i++) {
+        records_dto->pushBack(RandomBinRowRecordDto(dim));
     }
 
     return records_dto;
@@ -281,7 +307,7 @@ TEST_F(WebHandlerTest, INSERT_COUNT) {
     ASSERT_EQ(0, status_dto->code->getValue());
     ASSERT_EQ(1000, ids_dto->ids->count());
 
-    sleep(8);
+    sleep(2);
 
     milvus::server::web::OQueryParams query_params;
     query_params.put("fields", "num");
@@ -344,7 +370,7 @@ TEST_F(WebHandlerTest, PARTITION) {
     ASSERT_EQ(StatusCode::ILLEGAL_TABLE_NAME, status_dto->code->getValue());
 
     auto partitions_dto = milvus::server::web::PartitionListDto::createShared();
-    status_dto = handler->ShowPartitions(0, 10, table_name, partitions_dto);
+    status_dto = handler->ShowPartitions("0", "10", table_name, partitions_dto);
     ASSERT_EQ(1, partitions_dto->partitions->count());
 
     status_dto = handler->DropPartition(table_name, "test");
@@ -352,7 +378,7 @@ TEST_F(WebHandlerTest, PARTITION) {
 
     // Show all partitions
     partitions_dto = milvus::server::web::PartitionListDto::createShared();
-    status_dto = handler->ShowPartitions(0, 10, table_name, partitions_dto);
+    status_dto = handler->ShowPartitions("0", "10", table_name, partitions_dto);
 }
 
 TEST_F(WebHandlerTest, SEARCH) {
@@ -400,7 +426,54 @@ TEST_F(WebHandlerTest, CMD) {
 ///////////////////////////////////////////////////////////////////////////////////////
 
 namespace {
+static const char* CONTROLLER_TEST_VALID_CONFIG_STR =
+    "# Default values are used when you make no changes to the following parameters.\n"
+    "\n"
+    "version: 0.1"
+    "\n"
+    "server_config:\n"
+    "  address: 0.0.0.0                  # milvus server ip address (IPv4)\n"
+    "  port: 19530                       # port range: 1025 ~ 65534\n"
+    "  deploy_mode: single               \n"
+    "  time_zone: UTC+8\n"
+    "\n"
+    "db_config:\n"
+    "  backend_url: sqlite://:@:/        \n"
+    "\n"
+    "  insert_buffer_size: 4             # GB, maximum insert buffer size allowed\n"
+    "  preload_table:                    \n"
+    "\n"
+    "storage_config:\n"
+    "  primary_path: /tmp/milvus_web_controller_test        # path used to store data and meta\n"
+    "  secondary_path:                   # path used to store data only, split by semicolon\n"
+    "\n"
+    "metric_config:\n"
+    "  enable_monitor: false             # enable monitoring or not\n"
+    "  address: 127.0.0.1\n"
+    "  port: 8080                        # port prometheus uses to fetch metrics\n"
+    "\n"
+    "cache_config:\n"
+    "  cpu_cache_capacity: 4             # GB, CPU memory used for cache\n"
+    "  cpu_cache_threshold: 0.85         \n"
+    "  cache_insert_data: false          # whether to load inserted data into cache\n"
+    "\n"
+    "engine_config:\n"
+    "  use_blas_threshold: 20            \n"
+    "\n"
+    #ifdef MILVUS_GPU_VERSION
+    "gpu_resource_config:\n"
+    "  enable: true                      # whether to enable GPU resources\n"
+    "  cache_capacity: 4                 # GB, size of GPU memory per card used for cache, must be a positive integer\n"
+    "  search_resources:                 # define the GPU devices used for search computation, must be in format gpux\n"
+    "    - gpu0\n"
+    "  build_index_resources:            # define the GPU devices used for index building, must be in format gpux\n"
+    "    - gpu0\n"
+    #endif
+    "\n";
+
 static const char* CONTROLLER_TEST_TABLE_NAME = "controller_unit_test";
+static const char* CONTROLLER_TEST_CONFIG_DIR = "/tmp/milvus_web_controller_test/";
+static const char* CONTROLLER_TEST_CONFIG_FILE = "config.yaml";
 
 class TestClient : public oatpp::web::client::ApiClient {
  public:
@@ -445,11 +518,8 @@ class TestClient : public oatpp::web::client::ApiClient {
 
     API_CALL("OPTIONS", "/tables/{table_name}/indexes", optionsIndexes, PATH(String, table_name, "table_name"))
 
-    API_CALL("POST",
-             "/tables/{table_name}/indexes",
-             createIndex,
-             PATH(String, table_name, "table_name"),
-             BODY_DTO(milvus::server::web::IndexRequestDto::ObjectWrapper, body))
+    API_CALL("POST", "/tables/{table_name}/indexes",createIndex,
+             PATH(String, table_name, "table_name"), BODY_DTO(milvus::server::web::IndexRequestDto::ObjectWrapper, body))
 
     API_CALL("GET", "/tables/{table_name}/indexes", getIndex, PATH(String, table_name, "table_name"))
 
@@ -505,6 +575,15 @@ class WebControllerTest : public testing::Test {
  protected:
     static void
     SetUpTestCase() {
+        // Basic config
+        std::string config_path = std::string(CONTROLLER_TEST_CONFIG_DIR).append(CONTROLLER_TEST_CONFIG_FILE);
+        std::fstream fs(config_path.c_str(), std::ios_base::out);
+        fs << CONTROLLER_TEST_VALID_CONFIG_STR;
+        fs.close();
+
+        milvus::server::Config& config = milvus::server::Config::GetInstance();
+        config.LoadConfigFile(config_path);
+
         auto res_mgr = milvus::scheduler::ResMgrInst::GetInstance();
         res_mgr->Clear();
         res_mgr->Add(milvus::scheduler::ResourceFactory::Create("disk", "DISK", 0, false));
@@ -522,13 +601,8 @@ class WebControllerTest : public testing::Test {
         milvus::engine::DBOptions opt;
 
         milvus::server::Config::GetInstance().SetDBConfigBackendUrl("sqlite://:@:/");
-        boost::filesystem::remove_all("/tmp/milvus_web_controller_test");
-        milvus::server::Config::GetInstance().SetStorageConfigPrimaryPath("/tmp/milvus_web_controller_test");
-        milvus::server::Config::GetInstance().SetStorageConfigSecondaryPath("");
-        milvus::server::Config::GetInstance().SetDBConfigArchiveDiskThreshold("");
-        milvus::server::Config::GetInstance().SetDBConfigArchiveDaysThreshold("");
-        milvus::server::Config::GetInstance().SetCacheConfigCacheInsertData("");
-        milvus::server::Config::GetInstance().SetEngineConfigOmpThreadNum("");
+        boost::filesystem::remove_all(CONTROLLER_TEST_CONFIG_DIR);
+        milvus::server::Config::GetInstance().SetStorageConfigPrimaryPath(CONTROLLER_TEST_CONFIG_DIR);
 
         milvus::server::DBWrapper::GetInstance().StartService();
 
@@ -547,7 +621,7 @@ class WebControllerTest : public testing::Test {
         milvus::scheduler::JobMgrInst::GetInstance()->Stop();
         milvus::scheduler::ResMgrInst::GetInstance()->Stop();
         milvus::scheduler::SchedInst::GetInstance()->Stop();
-        boost::filesystem::remove_all("/tmp/milvus_web_controller_test");
+        boost::filesystem::remove_all(CONTROLLER_TEST_CONFIG_DIR);
     }
 
     void
@@ -638,16 +712,16 @@ TEST_F(WebControllerTest, CREATE_TABLE) {
     auto table_dto = milvus::server::web::TableRequestDto::createShared();
     auto response = client_ptr->createTable(table_dto, conncetion_ptr);
     ASSERT_EQ(OStatus::CODE_400.code, response->getStatusCode());
-    auto result_dto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
-    ASSERT_EQ(milvus::server::web::StatusCode::BODY_FIELD_LOSS, result_dto->code) << result_dto->message->std_str();
+    auto error_dto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
+    ASSERT_EQ(milvus::server::web::StatusCode::BODY_FIELD_LOSS, error_dto->code) << error_dto->message->std_str();
 
     OString table_name = "web_test_create_table" + OString(RandomName().c_str());
 
     table_dto->table_name = table_name;
     response = client_ptr->createTable(table_dto, conncetion_ptr);
     ASSERT_EQ(OStatus::CODE_400.code, response->getStatusCode());
-    result_dto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
-    ASSERT_EQ(milvus::server::web::StatusCode::BODY_FIELD_LOSS, result_dto->code) << result_dto->message->std_str();
+    error_dto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
+    ASSERT_EQ(milvus::server::web::StatusCode::BODY_FIELD_LOSS, error_dto->code) << error_dto->message->std_str();
 
     table_dto->dimension = 128;
     table_dto->index_file_size = 10;
@@ -655,6 +729,8 @@ TEST_F(WebControllerTest, CREATE_TABLE) {
 
     response = client_ptr->createTable(table_dto, conncetion_ptr);
     ASSERT_EQ(OStatus::CODE_201.code, response->getStatusCode());
+    auto result_dto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
+    ASSERT_EQ(milvus::server::web::StatusCode::SUCCESS, result_dto->code->getValue()) << result_dto->message->std_str();
 
     // invalid table name
     table_dto->table_name = "9090&*&()";
@@ -671,29 +747,32 @@ TEST_F(WebControllerTest, GET_TABLE) {
     // fields value is 'num', test count table
     params.put("fields", "num");
     auto response = client_ptr->getTable(table_name, conncetion_ptr);
-
     ASSERT_EQ(OStatus::CODE_200.code, response->getStatusCode());
     auto result_dto = response->readBodyToDto<milvus::server::web::TableFieldsDto>(object_mapper.get());
-
-    response = client_ptr->getTable(table_name, conncetion_ptr);
-    ASSERT_EQ(OStatus::CODE_200.code, response->getStatusCode());
+    ASSERT_EQ(table_name->std_str(), result_dto->table_name->std_str());
+    ASSERT_EQ(10, result_dto->dimension);
+    ASSERT_EQ("L2", result_dto->metric_type->std_str());
+    ASSERT_EQ(10, result_dto->index_file_size->getValue());
+    ASSERT_EQ("FLAT", result_dto->index->std_str());
 
     // invalid table name
     table_name = "57474dgdfhdfhdh  dgd";
     response = client_ptr->getTable(table_name, conncetion_ptr);
     ASSERT_EQ(OStatus::CODE_400.code, response->getStatusCode());
     auto status_sto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
+    ASSERT_EQ(milvus::server::web::StatusCode::ILLEGAL_TABLE_NAME, status_sto->code->getValue());
 
-    table_name = "test_table_not_found_0000000001110101010020202030203030435";
+    table_name = "test_table_not_found_000000000111010101002020203020aaaaa3030435";
     response = client_ptr->getTable(table_name, conncetion_ptr);
     ASSERT_EQ(OStatus::CODE_404.code, response->getStatusCode());
-    status_sto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
 }
 
 TEST_F(WebControllerTest, SHOW_TABLES) {
     // test query table limit 1
     auto response = client_ptr->showTables(1, 1, conncetion_ptr);
     ASSERT_EQ(OStatus::CODE_200.code, response->getStatusCode());
+    auto result_dto = response->readBodyToDto<milvus::server::web::TableListFieldsDto>(object_mapper.get());
+    ASSERT_TRUE(result_dto->count->getValue() > 0);
 
     // test query table empty
     response = client_ptr->showTables(0, 0, conncetion_ptr);
@@ -734,6 +813,24 @@ TEST_F(WebControllerTest, INSERT) {
     ASSERT_EQ(OStatus::CODE_204.code, response->getStatusCode());
 }
 
+TEST_F(WebControllerTest, INSERT_BIN) {
+    auto table_name = "test_insert_bin_table_test" + OString(RandomName().c_str());
+    const int64_t dim = 64;
+    GenTable(table_name, dim, 100, "HAMMING");
+
+    auto insert_dto = milvus::server::web::InsertRequestDto::createShared();
+    insert_dto->ids = insert_dto->ids->createShared();
+    insert_dto->records_bin = RandomBinRecordsDto(dim, 20);
+
+    auto response = client_ptr->insert(table_name, insert_dto, conncetion_ptr);
+    ASSERT_EQ(OStatus::CODE_201.code, response->getStatusCode());
+    auto result_dto = response->readBodyToDto<milvus::server::web::VectorIdsDto>(object_mapper.get());
+    ASSERT_EQ(20, result_dto->ids->count());
+
+    response = client_ptr->dropTable(table_name, conncetion_ptr);
+    ASSERT_EQ(OStatus::CODE_204.code, response->getStatusCode());
+}
+
 TEST_F(WebControllerTest, INSERT_IDS) {
     auto table_name = "test_insert_table_test" + OString(RandomName().c_str());
     const int64_t dim = 64;
@@ -764,6 +861,9 @@ TEST_F(WebControllerTest, INDEX) {
     auto index_dto = milvus::server::web::IndexRequestDto::createShared();
     auto response = client_ptr->createIndex(table_name, index_dto, conncetion_ptr);
     ASSERT_EQ(OStatus::CODE_201.code, response->getStatusCode());
+    auto create_index_dto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
+    ASSERT_EQ(milvus::server::web::StatusCode::SUCCESS, create_index_dto->code);
+
     // drop index
     response = client_ptr->dropIndex(table_name, conncetion_ptr);
     ASSERT_EQ(OStatus::CODE_204.code, response->getStatusCode());
@@ -795,7 +895,6 @@ TEST_F(WebControllerTest, INDEX) {
     // invalid index type
     index_dto->index_type = 100;
     response = client_ptr->createIndex(table_name, index_dto, conncetion_ptr);
-    ASSERT_NE(OStatus::CODE_201.code, response->getStatusCode());
     ASSERT_EQ(OStatus::CODE_400.code, response->getStatusCode());
 
     // insert data and create index
@@ -816,6 +915,9 @@ TEST_F(WebControllerTest, INDEX) {
     // get index
     response = client_ptr->getIndex(table_name, conncetion_ptr);
     ASSERT_EQ(OStatus::CODE_200.code, response->getStatusCode());
+    auto result_index_dto = response->readBodyToDto<milvus::server::web::IndexDto>(object_mapper.get());
+    ASSERT_EQ("FLAT", result_index_dto->index_type->std_str());
+    ASSERT_EQ(10, result_index_dto->nlist->getValue());
 }
 
 TEST_F(WebControllerTest, PARTITION) {
@@ -825,23 +927,25 @@ TEST_F(WebControllerTest, PARTITION) {
     auto par_param = milvus::server::web::PartitionRequestDto::createShared();
     auto response = client_ptr->createPartition(table_name, par_param);
     ASSERT_EQ(OStatus::CODE_400.code, response->getStatusCode());
-    auto result_dto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
-    ASSERT_EQ(milvus::server::web::StatusCode::BODY_FIELD_LOSS, result_dto->code);
+    auto error_dto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
+    ASSERT_EQ(milvus::server::web::StatusCode::BODY_FIELD_LOSS, error_dto->code);
 
     par_param->partition_name = "partition01" + OString(RandomName().c_str());
     response = client_ptr->createPartition(table_name, par_param);
-    result_dto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
-    ASSERT_EQ(milvus::server::web::StatusCode::BODY_FIELD_LOSS, result_dto->code);
+    ASSERT_EQ(OStatus::CODE_400.code, response->getStatusCode());
+    error_dto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
+    ASSERT_EQ(milvus::server::web::StatusCode::BODY_FIELD_LOSS, error_dto->code);
 
     par_param->partition_tag = "tag01";
     response = client_ptr->createPartition(table_name, par_param);
     ASSERT_EQ(OStatus::CODE_201.code, response->getStatusCode());
+    auto create_result_dto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
+    ASSERT_EQ(milvus::server::web::StatusCode::SUCCESS, create_result_dto->code);
 
     // insert 200 vectors into table with tag = 'tag01'
     OQueryParams query_params;
     // add partition tag
     auto insert_dto = milvus::server::web::InsertRequestDto::createShared();
-    // add partition tag
     insert_dto->tag = OString("tag01");
     insert_dto->ids = insert_dto->ids->createShared();
     insert_dto->records = insert_dto->records->createShared();
@@ -854,13 +958,17 @@ TEST_F(WebControllerTest, PARTITION) {
     // Show all partitins
     response = client_ptr->showPartitions(table_name, 0, 10, conncetion_ptr);
     ASSERT_EQ(OStatus::CODE_200.code, response->getStatusCode());
+    auto result_dto = response->readBodyToDto<milvus::server::web::PartitionListDto>(object_mapper.get());
+    ASSERT_EQ(1, result_dto->partitions->count());
+    ASSERT_EQ("tag01", result_dto->partitions->get(0)->partition_tag->std_str());
+    ASSERT_EQ(par_param->partition_name->std_str(), result_dto->partitions->get(0)->partition_name->std_str());
 
     response = client_ptr->dropPartition(table_name, "tag01", conncetion_ptr);
     ASSERT_EQ(OStatus::CODE_204.code, response->getStatusCode());
 }
 
 TEST_F(WebControllerTest, SEARCH) {
-    const OString table_name = "test_partition_table_test" + OString(RandomName().c_str());
+    const OString table_name = "test_search_table_test" + OString(RandomName().c_str());
     GenTable(table_name, 64, 100, "L2");
 
     // Insert 200 vectors into table
@@ -868,6 +976,67 @@ TEST_F(WebControllerTest, SEARCH) {
     auto insert_dto = milvus::server::web::InsertRequestDto::createShared();
     insert_dto->ids = insert_dto->ids->createShared();
     insert_dto->records = RandomRecordsDto(64, 200);// insert_dto->records->createShared();
+
+    auto response = client_ptr->insert(table_name, insert_dto, conncetion_ptr);
+    ASSERT_EQ(OStatus::CODE_201.code, response->getStatusCode());
+    auto insert_result_dto = response->readBodyToDto<milvus::server::web::VectorIdsDto>(object_mapper.get());
+    ASSERT_EQ(200, insert_result_dto->ids->count());
+
+    sleep(4);
+
+    //Create partition and insert 200 vectors into it
+    auto par_param = milvus::server::web::PartitionRequestDto::createShared();
+    par_param->partition_name = "partition" + OString(RandomName().c_str());
+    par_param->partition_tag = "tag" + OString(RandomName().c_str());
+    response = client_ptr->createPartition(table_name, par_param);
+    ASSERT_EQ(OStatus::CODE_201.code, response->getStatusCode())
+                        << "Error: " << response->getStatusDescription()->std_str();
+
+    insert_dto->tag = par_param->partition_tag;
+    response = client_ptr->insert(table_name, insert_dto, conncetion_ptr);
+    ASSERT_EQ(OStatus::CODE_201.code, response->getStatusCode());
+    sleep(2);
+
+    // Test search
+    auto search_request_dto = milvus::server::web::SearchRequestDto::createShared();
+    response = client_ptr->search(table_name, search_request_dto, conncetion_ptr);
+    auto error_dto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
+    ASSERT_EQ(milvus::server::web::StatusCode::BODY_FIELD_LOSS, error_dto->code);
+
+    search_request_dto->nprobe = 1;
+    response = client_ptr->search(table_name, search_request_dto, conncetion_ptr);
+    error_dto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
+    ASSERT_EQ(milvus::server::web::StatusCode::BODY_FIELD_LOSS, error_dto->code);
+
+    search_request_dto->topk = 1;
+    response = client_ptr->search(table_name, search_request_dto, conncetion_ptr);
+    error_dto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
+    ASSERT_EQ(milvus::server::web::StatusCode::BODY_FIELD_LOSS, error_dto->code);
+
+    search_request_dto->records = RandomRecordsDto(64, 10);
+    response = client_ptr->search(table_name, search_request_dto, conncetion_ptr);
+    ASSERT_EQ(OStatus::CODE_200.code, response->getStatusCode());
+    auto result_dto = response->readBodyToDto<milvus::server::web::TopkResultsDto>(object_mapper.get());
+    ASSERT_EQ(10, result_dto->num);
+    ASSERT_EQ(10, result_dto->results->count());
+    ASSERT_EQ(1, result_dto->results->get(0)->count());
+
+    // Test search with tags
+    search_request_dto->tags = search_request_dto->tags->createShared();
+    search_request_dto->tags->pushBack(par_param->partition_tag);
+    response = client_ptr->search(table_name, search_request_dto, conncetion_ptr);
+    ASSERT_EQ(OStatus::CODE_200.code, response->getStatusCode());
+}
+
+TEST_F(WebControllerTest, SEARCH_BIN) {
+    const OString table_name = "test_search_bin_table_test" + OString(RandomName().c_str());
+    GenTable(table_name, 64, 100, "HAMMING");
+
+    // Insert 200 vectors into table
+    OQueryParams query_params;
+    auto insert_dto = milvus::server::web::InsertRequestDto::createShared();
+    insert_dto->ids = insert_dto->ids->createShared();
+    insert_dto->records_bin = RandomBinRecordsDto(64, 200);
 
     auto response = client_ptr->insert(table_name, insert_dto, conncetion_ptr);
     ASSERT_EQ(OStatus::CODE_201.code, response->getStatusCode());
@@ -903,7 +1072,7 @@ TEST_F(WebControllerTest, SEARCH) {
     result_dto = response->readBodyToDto<milvus::server::web::StatusDto>(object_mapper.get());
     ASSERT_EQ(milvus::server::web::StatusCode::BODY_FIELD_LOSS, result_dto->code);
 
-    search_request_dto->records = RandomRecordsDto(64, 10);
+    search_request_dto->records_bin = RandomBinRecordsDto(64, 10);
     response = client_ptr->search(table_name, search_request_dto, conncetion_ptr);
     ASSERT_EQ(OStatus::CODE_200.code, response->getStatusCode());