+++ CHANGELOG.md
@@ -11,6 +11,13 @@ Please mark all change in change log and use the issue from GitHub
 - \#805 - IVFTest.gpu_seal_test unittest failed
 - \#831 - Judge branch error in CommonUtil.cpp
 - \#977 - Server crash when create tables concurrently
+- \#995 - table count set to 0 if no tables found
+- \#1010 - improve error message when offset or page_size is equal 0
+- \#1022 - check if partition name is legal
+- \#1028 - check if table exists when show partitions
+- \#1029 - check if table exists when try to delete partition
+- \#1066 - optimize http insert and search speed
+- \#1067 - Add binary vectors support in http server
 
 ## Feature
 - \#216 - Add CLI to get server info
@@ -35,6 +42,8 @@ Please mark all change in change log and use the issue from GitHub
 - \#908 - Move "primary_path" and "secondary_path" to storage config
 - \#931 - Remove "collector" from config
 - \#966 - Update NOTICE.md
+- \#1002 - Rename minio to s3 in Storage Config section
+- \#1078 - Move 'insert_buffer_size' to Cache Config section
 
 ## Task